+++ lib/ecto_shorts/common_filters/having.ex
@@ -1,5 +1,5 @@
 defmodule EctoShorts.CommonFilters.Having do
-  alias EctoShorts.Dynamics.Adapters.Postgres
+  alias EctoShorts.Dynamics.Postgres
   alias EctoShorts.QueryBindings
 
   alias Ecto.Query